ALL: add VTX bit support, correct 'ls -la' output to show t,T,+ to indicate VTX, VTX+ACL, ACL

+++ mgm/XrdMgmOfs/Remdir.cc
@@ -104,10 +104,14 @@ XrdMgmOfs::_remdir(const char* path,
   eos::common::RWMutexWriteLock viewLock(gOFS->eosViewRWMutex);
   std::string aclpath;
 
+  uid_t container_owner_uid=0;
+  bool container_vtx = false;
   try {
     dh = gOFS->eosView->getContainer(path);
     eos::common::Path pPath(gOFS->eosView->getUri(dh.get()).c_str());
     dhpar = gOFS->eosView->getContainer(pPath.GetParentPath());
+    container_owner_uid=dh->getCUid();
+    container_vtx = dhpar->getMode() & S_ISVTX;
     aclpath = pPath.GetParentPath();
   } catch (eos::MDException& e) {
     dh.reset();
@@ -119,30 +123,30 @@ XrdMgmOfs::_remdir(const char* path,
 
   eos_info("path='%s' scope='%s' aclpath='%s'\n", path, vid.scope.c_str(),
            aclpath.c_str());
-
+  
   // check existence
   if (!dh) {
     errno = ENOENT;
     return Emsg(epname, error, errno, "rmdir", path);
   }
-
+  
   // ACL and permission check
   Acl acl(aclpath.c_str(), error, vid, attrmap, false);
-
+  
   if (vid.uid && !acl.IsMutable()) {
     errno = EPERM;
     return Emsg(epname, error, EPERM, "rmdir - immutable", path);
   }
-
+  
   if (!gOFS->allow_public_access(aclpath.c_str(), vid)) {
     errno = EACCES;
     return Emsg(epname, error, EACCES, "access - public access level restriction",
                 aclpath.c_str());
   }
-
+  
   if (ininfo) {
     XrdOucEnv env_info(ininfo);
-
+    
     if (env_info.Get("mgm.option")) {
       XrdOucString option = env_info.Get("mgm.option");
 
@@ -167,18 +171,20 @@ XrdMgmOfs::_remdir(const char* path,
 
   bool stdpermcheck = false;
   bool aclok = false;
-
-  if (acl.HasAcl()) {
+  
+  if (acl.HasAcl() && !container_vtx) {
+    // acls only if this is not a VTX directory
     if ((dh->getCUid() != vid.uid) &&
         (vid.uid) && // not the root user
         (vid.uid != 3) && // not the admin user
         (vid.gid != 4) && // not the admin group
-        (acl.CanNotDelete())) {
-      // deletion is explicitly forbidden
-      errno = EPERM;
-      return Emsg(epname, error, EPERM, "rmdir by ACL", path);
-    }
-
+        (acl.CanNotDelete()))  // acl does not allow deletion
+      {
+	// deletion is explicitly forbidden
+	errno = EPERM;
+	return Emsg(epname, error, EPERM, "rmdir by ACL", path);
+      }
+    
     if ((!acl.CanWrite())) {
       // we have to check the standard permissions
       stdpermcheck = true;
@@ -191,25 +197,34 @@ XrdMgmOfs::_remdir(const char* path,
 
   // Check permissions
   bool permok = stdpermcheck ? (dhpar ? (dhpar->access(vid.uid, vid.gid,
-                                         X_OK | W_OK)) : false) : aclok;
-
-  if (!permok) {
-    errno = EPERM;
-    return Emsg(epname, error, errno, "rmdir", path);
+						       X_OK | W_OK)) : false) : aclok;
+  
+  if (container_vtx) {
+    if (container_owner_uid != vid.uid) {
+      // only the owner can delete
+      errno = EPERM;
+      return Emsg(epname, error, errno, "rmdir", path);
+    }
+  } else {
+    // need for standard perm check
+    if (!permok) {
+      errno = EPERM;
+      return Emsg(epname, error, errno, "rmdir", path);
+    }
   }
-
+  
   if ((dh->getFlags() && eos::QUOTA_NODE_FLAG) && (vid.uid)) {
     errno = EADDRINUSE;
     eos_err("%s is a quota node - deletion canceled", path);
     return Emsg(epname, error, errno, "rmdir - this is a quota node", path);
   }
-
+  
   if (!simulate) {
     try {
       eos::ContainerIdentifier dhpar_id;
       eos::ContainerIdentifier dhpar_pid;
       std::string dh_name;
-
+      
       // update the in-memory modification time of the parent directory
       if (dhpar) {
         dhpar->setMTimeNow();
@@ -218,7 +233,7 @@ XrdMgmOfs::_remdir(const char* path,
         dhpar_id = dhpar->getIdentifier();
         dh_name = dh->getName();
       }
-
+      
       eosView->removeContainer(path);
       viewLock.Release();
